refactor : api 구조 수정
