+++ src/api/services/gallery.js
@@ -1,0 +1,19 @@
+import * as EgovNet from "@/api/egovFetch";
+
+export const getGalleryList = async (searchCondition) => {
+  const retrieveListURL = "/board" + EgovNet.getQueryString(searchCondition);
+  const requestOptions = {
+    method: "GET",
+    headers: {
+      "Content-type": "application/json",
+    },
+  };
+  return new Promise((resolve, reject) => {
+    EgovNet.requestFetch(
+      retrieveListURL,
+      requestOptions,
+      (resp) => resolve(resp),
+      (err) => reject(err)
+    );
+  });
+};
