+++ src/pages/inform/gallery/EgovGalleryList.jsx
@@ -1,7 +1,6 @@
-import { useState, useEffect, useCallback, useRef } from "react";
+import { useState, useEffect, useRef } from "react";
 import { Link, useLocation } from "react-router-dom";
 
-import * as EgovNet from "@/api/egovFetch";
 import URL from "@/constants/url";
 import { GALLERY_BBS_ID } from "@/config";
 
@@ -9,6 +8,7 @@ import { default as EgovLeftNav } from "@/components/leftmenu/EgovLeftNavInform"
 import EgovPaging from "@/components/EgovPaging";
 
 import { itemIdxByPage } from "@/utils/calc";
+import { getGalleryList } from "@/api/services/gallery";
 
 function EgovGalleryList(props) {
   console.group("EgovGalleryList");
@@ -38,87 +38,68 @@ function EgovGalleryList(props) {
 
   const [listTag, setListTag] = useState([]);
 
-  const retrieveList = useCallback((searchCondition) => {
+  const retrieveList = async (searchCondition) => {
     console.groupCollapsed("EgovGalleryList.retrieveList()");
 
-    const retrieveListURL = "/board" + EgovNet.getQueryString(searchCondition);
-    const requestOptions = {
-      method: "GET",
-      headers: {
-        "Content-type": "application/json",
-      },
-    };
+    const resp = await getGalleryList(searchCondition);
 
-    EgovNet.requestFetch(
-      retrieveListURL,
-      requestOptions,
-      (resp) => {
-        setMasterBoard(resp.result.brdMstrVO);
-        setPaginationInfo(resp.result.paginationInfo);
-        setUser(resp.result.user);
+    console.log("RESP", console.log(JSON.stringify(resp)));
+    setMasterBoard(resp.result.brdMstrVO);
+    setPaginationInfo(resp.result.paginationInfo);
+    setUser(resp.result.user);
 
-        let mutListTag = [];
-        mutListTag.push(
-          <p className="no_data" key="0">
-            검색된 결과가 없습니다.
-          </p>
-        ); // 게시판 목록 초기값
+    let mutListTag = [];
+    mutListTag.push(
+      <p className="no_data" key="0">
+        검색된 결과가 없습니다.
+      </p>
+    ); // 게시판 목록 초기값
 
-        const resultCnt = parseInt(resp.result.resultCnt);
-        const currentPageNo = resp.result.paginationInfo.currentPageNo;
-        const pageSize = resp.result.paginationInfo.pageSize;
+    const resultCnt = parseInt(resp.result.resultCnt);
+    const currentPageNo = resp.result.paginationInfo.currentPageNo;
+    const pageSize = resp.result.paginationInfo.pageSize;
 
-        // 리스트 항목 구성
-        resp.result.resultList.forEach(function (item, index) {
-          if (index === 0) mutListTag = []; // 목록 초기화
-          const listIdx = itemIdxByPage(
-            resultCnt,
-            currentPageNo,
-            pageSize,
-            index
-          );
+    // 리스트 항목 구성
+    resp.result.resultList.forEach(function (item, index) {
+      if (index === 0) mutListTag = []; // 목록 초기화
+      const listIdx = itemIdxByPage(resultCnt, currentPageNo, pageSize, index);
 
-          mutListTag.push(
-            <Link
-              to={{ pathname: URL.INFORM_GALLERY_DETAIL }}
-              state={{
-                nttId: item.nttId,
-                bbsId: item.bbsId,
-                searchCondition: searchCondition,
-              }}
-              key={listIdx}
-              className="list_item"
-            >
-              <div>{listIdx}</div>
-              {(item.replyLc * 1 ? true : false) && (
-                <>
-                  <div className="al reply">{item.nttSj}</div>
-                </>
-              )}
-              {(item.replyLc * 1 ? false : true) && (
-                <>
-                  <div className="al">{item.nttSj}</div>
-                </>
-              )}
-              <div>{item.frstRegisterNm}</div>
-              <div>{item.frstRegisterPnttm}</div>
-              <div>{item.inqireCo}</div>
-            </Link>
-          );
-        });
-        setListTag(mutListTag);
-      },
-      function (resp) {
-        console.log("err response : ", resp);
-      }
-    );
+      mutListTag.push(
+        <Link
+          to={{ pathname: URL.INFORM_GALLERY_DETAIL }}
+          state={{
+            nttId: item.nttId,
+            bbsId: item.bbsId,
+            searchCondition: searchCondition,
+          }}
+          key={listIdx}
+          className="list_item"
+        >
+          <div>{listIdx}</div>
+          {(item.replyLc * 1 ? true : false) && (
+            <>
+              <div className="al reply">{item.nttSj}</div>
+            </>
+          )}
+          {(item.replyLc * 1 ? false : true) && (
+            <>
+              <div className="al">{item.nttSj}</div>
+            </>
+          )}
+          <div>{item.frstRegisterNm}</div>
+          <div>{item.frstRegisterPnttm}</div>
+          <div>{item.inqireCo}</div>
+        </Link>
+      );
+    });
+    setListTag(mutListTag);
+
     console.groupEnd("EgovGalleryList.retrieveList()");
-  }, []);
+  };
 
   //======================================================
   useEffect(() => {
     retrieveList(searchCondition);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   console.log("------------------------------EgovGalleryList [End]");
